requirements.txt
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
et_xmlfile==2.0.0
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pycparser==2.23
pypdfium2==5.0.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
